separando en distintos archivos a el controlador y el routeador para mejor organizacion
