+++ src/controllers/loginController.js
@@ -1,0 +1,13 @@
+const path = require("path");
+
+
+const loginControlador={
+    login: (req, res) => {
+        res.render("login");
+    },
+    guardarlogin: (req, res) => {
+        return res.redirect("/")
+      },
+}
+
+module.exports = loginControlador;
